mcp/semantic_search_mcp/requirements.txt
fastapi
uvicorn[standard]
weaviate-client[agents]
pydantic
